rei_models/requirements.txt
pydantic[email]~=2.3.0
pytest~=8.3.2
python-dateutil~=2.9.0.post0